Ajustes na tela de profile

+++ src/app/(tabs)/profile.tsx
@@ -1,17 +1,17 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, ScrollView, StyleSheet, ImageBackground, TouchableOpacity, Alert } from 'react-native';
-import { useForm, Controller, SubmitHandler } from 'react-hook-form';
-import { globalStyles } from '@/styles/styles';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import { useAuthContext } from '@/src/contexts/AuthContext';
-import { colors } from '@/styles/color';
-import { Input } from '@/src/components/Input';
-import Button from '@/src/components/Button';
-import { getPeople, updatePeople } from '@/src/httpservices/user';
-import { useIsFocused } from '@react-navigation/native';
-import { IMAGE_PATHS } from '@/styles/constants';
-import { Ionicons } from '@expo/vector-icons';
-import { useSQLiteContext } from 'expo-sqlite';
+import React, { useState, useEffect } from "react";
+import { View, Text, ScrollView, StyleSheet, ImageBackground, TouchableOpacity, Alert } from "react-native";
+import { useForm, Controller, SubmitHandler } from "react-hook-form";
+import { globalStyles } from "@/styles/styles";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { useAuthContext } from "@/src/contexts/AuthContext";
+import { colors } from "@/styles/color";
+import { Input } from "@/src/components/Input";
+import Button from "@/src/components/Button";
+import { getPeople, updatePeople } from "@/src/httpservices/user";
+import { useIsFocused } from "@react-navigation/native";
+import { IMAGE_PATHS } from "@/styles/constants";
+import { Ionicons } from "@expo/vector-icons";
+import { useSQLiteContext } from "expo-sqlite";
 
 export type FormType = {
     email: string;
@@ -24,11 +24,10 @@ export default function ProfileForm() {
     const { control, handleSubmit, watch, reset, formState: { errors } } = useForm<FormType>();
     const { signOut } = useAuthContext();
     const { user } = useAuthContext();
-    const [error, setError] = useState<string>('');
-    const password = watch('password');
+    const [error, setError] = useState<string>("");
+    const password = watch("password");
     const isFocused = useIsFocused();
     const database = useSQLiteContext();
-    const [erro, setErro] = useState<string>('');
 
     const [editStates, setEditStates] = useState({
         name: false,
@@ -44,25 +43,27 @@ export default function ProfileForm() {
                     reset({
                         name: response.data.items[0].nome,
                         email: response.data.items[0].email,
-                        password: '',
-                        confirmPassword: '',
+                        password: "",
+                        confirmPassword: "",
                     });
                 } else {
-                    setError('Erro ao carregar informações do perfil.');
+                    setError("Erro ao carregar informações do perfil.");
                 }
             } catch (e) {
-                setError('Falha na conexão.');
+                setError("Falha na conexão.");
             }
         }
 
         if (isFocused) {
             fetchUserProfile();
+        } else {
+            setError("");
         }
     }, [isFocused, user?.id_pessoa, reset]);
 
     const onSubmit: SubmitHandler<FormType> = async (data) => {
-        setError('');
         try {
+            setError("")
             if (user?.id_pessoa) {
                 const response = await updatePeople({
                     id_pessoa: user?.id_pessoa,
@@ -74,34 +75,46 @@ export default function ProfileForm() {
 
                 if (response?.status === 555) {
                     if (response.data.cause.includes("ORA-00001")) {
-                        setErro("Email já cadastrado");
+                        setError("Email já cadastrado");
                     } else if (response.data.cause.includes("ORA-02291")) {
-                        setErro("Falha na conexão: Tipo de perfil não encontrado");
+                        setError("Falha na conexão: Tipo de perfil não encontrado");
                     }
                 } else if (response?.status === 571) {
-                    setErro("Falha na conexão");
+                    setError("Falha na conexão");
+                } else {
+                    setError("")
+                    setEditStates({
+                        name: false,
+                        email: false,
+                        password: false,
+                    })
                 }
-                setEditStates({
-                    name: false,
-                    email: false,
-                    password: false,
-                });
             }
         } catch (e) {
-            setError('Falha na conexão.');
+            setError("Falha na conexão.");
         }
     };
 
 
     async function logout() {
-        const tables: { name: string }[] = await database.getAllAsync(`SELECT name FROM sqlite_master WHERE type='table'`);
-        for (const table of tables) {
-            await database.execAsync(`DELETE FROM ${table.name}`);
-        }
-        signOut();
+        Alert.alert("Sair?", "Deseja sair da sua conta?", [
+            {
+                text: "Cancelar"
+            },
+            {
+                text: "Confirmar",
+                onPress: async () => {
+                    const tables: { name: string }[] = await database.getAllAsync(`SELECT name FROM sqlite_master WHERE type="table"`);
+                    for (const table of tables) {
+                        await database.execAsync(`DELETE FROM ${table.name}`);
+                    }
+                    signOut();
+                }
+            }
+        ])
     }
 
-    const toggleEdit = (field: 'name' | 'email' | 'password') => {
+    const toggleEdit = (field: "name" | "email" | "password") => {
         setEditStates((prevState) => ({
             ...prevState,
             [field]: !prevState[field],
@@ -111,7 +124,7 @@ export default function ProfileForm() {
     return (
         <ImageBackground source={IMAGE_PATHS.backgroundImage} style={globalStyles.backgroundImage}>
             <SafeAreaView style={[globalStyles.pageContainer, { flex: 1, paddingBottom: 0 }]}>
-                <ScrollView style={{ flexGrow: 0, width: '100%' }}>
+                <ScrollView style={{ flexGrow: 0, width: "100%" }}>
                     <View style={globalStyles.container}>
                         <View style={{ flexDirection: "row", alignItems: "center", justifyContent: "space-between" }}>
                             <Text style={[globalStyles.title, { color: colors.primary }]}>
@@ -121,7 +134,7 @@ export default function ProfileForm() {
                                 <Ionicons name={"exit-outline"} size={30} color={colors.error} />
                             </TouchableOpacity>
                         </View>
-                        {erro && <Text style={{ color: error === "Dados atualizados" ? colors.primary : colors.error }}>{erro}</Text>}
+                        {error && <Text style={{ color: error === "Dados atualizados" ? colors.primary : colors.error }}>{error}</Text>}
 
                         <View style={globalStyles.formContainer}>
                             {error && <Text style={{ color: colors.error }}>{error}</Text>}
@@ -131,7 +144,7 @@ export default function ProfileForm() {
                                     <Controller
                                         control={control}
                                         name="name"
-                                        rules={{ required: 'Nome é obrigatório' }}
+                                        rules={{ required: "Nome é obrigatório" }}
                                         render={({ field: { onChange, value } }) => (
                                             <Input
                                                 label="Nome"
@@ -145,7 +158,7 @@ export default function ProfileForm() {
                                 ) : (
                                     <Text style={styles.text}>Nome: {watch("name")}</Text>
                                 )}
-                                <TouchableOpacity onPress={() => toggleEdit('name')}>
+                                <TouchableOpacity onPress={() => toggleEdit("name")}>
                                     <Ionicons name={editStates.name ? "arrow-back-outline" : "create-outline"} size={30} color={colors.primary} />
                                 </TouchableOpacity>
                             </View>
@@ -157,10 +170,10 @@ export default function ProfileForm() {
                                         control={control}
                                         name="email"
                                         rules={{
-                                            required: 'Email é obrigatório',
+                                            required: "Email é obrigatório",
                                             pattern: {
                                                 value: /\S+@\S+\.\S+/,
-                                                message: 'Formato de email inválido',
+                                                message: "Formato de email inválido",
                                             },
                                         }}
                                         render={({ field: { onChange, value } }) => (
@@ -179,7 +192,7 @@ export default function ProfileForm() {
                                 ) : (
                                     <Text style={styles.text}>Email: {watch("email")}</Text>
                                 )}
-                                <TouchableOpacity onPress={() => toggleEdit('email')}>
+                                <TouchableOpacity onPress={() => toggleEdit("email")}>
                                     <Ionicons name={editStates.email ? "arrow-back-outline" : "create-outline"} size={30} color={colors.primary} />
                                 </TouchableOpacity>
                             </View>
@@ -190,7 +203,7 @@ export default function ProfileForm() {
                                     <Controller
                                         control={control}
                                         name="password"
-                                        rules={{ required: 'Senha é obrigatória' }}
+                                        rules={{ required: "Senha é obrigatória" }}
                                         render={({ field: { onChange, value } }) => (
                                             <Input
                                                 label="Senha"
@@ -206,7 +219,7 @@ export default function ProfileForm() {
                                 ) : (
                                     <Text style={styles.text}>Senha: ******</Text>
                                 )}
-                                <TouchableOpacity onPress={() => toggleEdit('password')}>
+                                <TouchableOpacity onPress={() => toggleEdit("password")}>
                                     <Ionicons name={editStates.password ? "arrow-back-outline" : "create-outline"} size={30} color={colors.primary} />
                                 </TouchableOpacity>
                             </View>
@@ -225,9 +238,9 @@ export default function ProfileForm() {
 
 const styles = StyleSheet.create({
     inputContainer: {
-        flexDirection: 'row',
+        flexDirection: "row",
         alignItems: "center",
-        justifyContent: 'space-between',
+        justifyContent: "space-between",
         gap: 8,
         marginVertical: 10,
     },
